feat: fetch polkadot-omni-node with pop up command
