+++ crates/pop-cli/src/common/mod.rs
@@ -12,6 +12,7 @@ pub mod chain;
 pub mod contracts;
 #[cfg(any(feature = "chain", feature = "wasm-contracts", feature = "polkavm-contracts"))]
 pub mod helpers;
+/// Contains omni-node utilities.
 #[cfg(feature = "chain")]
 pub mod omni_node;
 /// Contains utilities for interacting with the CLI prompt.
